[refactor] 계정 소개글 조회 API 이동
