+++ apps/backend/src/auth/auth.queries.ts
@@ -4,5 +4,6 @@ export const authQueries = {
     'SELECT member_id, email, password, nickname FROM members WHERE email = $1 LIMIT 1',
   updateInroduceQuery: 'UPDATE members SET introduce = $1 WHERE member_id = $2',
   updateNicknameQuery: 'UPDATE members SET nickname = $1 WHERE member_id = $2',
-  createLottoColumnQuery: 'INSERT INTO lottos (member_id) VALUES ($1)'
+  createLottoColumnQuery: 'INSERT INTO lottos (member_id) VALUES ($1)',
+  getIntroduceQuery: 'SELECT introduce FROM members WHERE member_id = $1'
 };
